update: ui of info
