+++ src/components/info/GroupInfo.tsx
@@ -3,6 +3,7 @@ import { View, Text, Image, TouchableOpacity, Modal, TextInput, ScrollView } fro
 import { Ionicons } from "@expo/vector-icons";
 import { Conversation } from "@/src/models/Conversation";
 import { Dimensions } from "react-native";
+import AddMemberModal from "./AddMemberModal";// Update the path to the correct location of AddMemberModal
 
 interface GroupInfoProps {
   group: Conversation;
@@ -101,11 +102,6 @@ const MOCK_MEMBERS = [
   },
 ];
 
-interface AddMemberModalProps {
-  visible: boolean;
-  onClose: () => void;
-}
-
 interface MemberMenuProps {
   visible: boolean;
   onClose: () => void;
@@ -113,123 +109,6 @@ interface MemberMenuProps {
 }
 
 const isDesktop = Dimensions.get('window').width > 768;
-
-const AddMemberModal = ({ visible, onClose }: AddMemberModalProps) => {
-  const [searchQuery, setSearchQuery] = useState('');
-  const [selectedUsers, setSelectedUsers] = useState<string[]>([]);
-
-  const toggleUserSelection = (userId: string) => {
-    setSelectedUsers(prev =>
-      prev.includes(userId)
-        ? prev.filter(id => id !== userId)
-        : [...prev, userId]
-    );
-  };
-
-  const filteredUsers = MOCK_USERS.filter(user =>
-    user.name.toLowerCase().includes(searchQuery.toLowerCase())
-  );
-
-  return (
-    <Modal
-      visible={visible}
-      animationType="slide"
-      transparent={true}
-    >
-      <View className="flex-1 bg-black/30 items-center">
-        {/* Container cho desktop */}
-        <View style={[
-          {
-            width: isDesktop ? 400 : '100%',
-            height: isDesktop ? 480 : '100%',
-            borderRadius: isDesktop ? 16 : 0,
-            overflow: 'hidden',
-            marginTop: isDesktop ? 100 : 0
-          }
-        ]}>
-          {/* Content container */}
-          <View style={[
-            {
-              flex: 1,
-              backgroundColor: 'white',
-              marginTop: isDesktop ? 0 : 64,
-              borderTopLeftRadius: isDesktop ? 0 : 24,
-              borderTopRightRadius: isDesktop ? 0 : 24
-            }
-          ]}>
-            {/* Header */}
-            <View className="flex-row items-center justify-between px-4 py-3 border-b border-gray-200">
-              <TouchableOpacity onPress={onClose} className="p-2 -ml-2">
-                <Ionicons name="close" size={24} color="#666" />
-              </TouchableOpacity>
-              <Text className="text-lg font-semibold">Thêm thành viên</Text>
-              <TouchableOpacity
-                className={`py-1 px-3 rounded-lg ${selectedUsers.length > 0 ? 'bg-blue-500' : 'bg-gray-200'}`}
-                disabled={selectedUsers.length === 0}
-              >
-                <Text className={selectedUsers.length > 0 ? 'text-white' : 'text-gray-500'}>
-                  Thêm ({selectedUsers.length})
-                </Text>
-              </TouchableOpacity>
-            </View>
-
-            {/* Search Bar */}
-            <View className="px-4 py-2 border-b border-gray-200">
-              <View className="flex-row items-center bg-gray-100 rounded-full px-4 py-2">
-                <Ionicons name="search" size={20} color="#666" />
-                <TextInput
-                  className="flex-1 ml-2 text-base"
-                  placeholder="Tìm kiếm bạn bè"
-                  placeholderTextColor="#666"
-                  value={searchQuery}
-                  onChangeText={setSearchQuery}
-                />
-              </View>
-            </View>
-
-            {/* User List */}
-            <ScrollView className="flex-1 px-4">
-              {filteredUsers.map(user => (
-                <TouchableOpacity
-                  key={user.id}
-                  className="flex-row items-center py-3 hover:bg-gray-50 active:bg-gray-100"
-                  onPress={() => toggleUserSelection(user.id)}
-                >
-                  <View className="relative">
-                    <Image
-                      source={{ uri: user.avatar }}
-                      className="w-12 h-12 rounded-full"
-                    />
-                    {user.isOnline && (
-                      <View className="absolute bottom-0 right-0 w-3 h-3 rounded-full bg-emerald-500 border-2 border-white" />
-                    )}
-                  </View>
-                  <View className="flex-1 ml-3">
-                    <Text className="text-[15px] font-medium text-gray-900">{user.name}</Text>
-                    <Text className="text-sm text-gray-500">
-                      {user.mutualFriends} bạn chung
-                    </Text>
-                  </View>
-                  <View className={`w-6 h-6 rounded-full border-2 items-center justify-center
-                    ${selectedUsers.includes(user.id)
-                      ? 'bg-blue-500 border-blue-500'
-                      : 'border-gray-300'}`}
-                  >
-                    {selectedUsers.includes(user.id) && (
-                      <Ionicons name="checkmark" size={16} color="white" />
-                    )}
-                  </View>
-                </TouchableOpacity>
-              ))}
-              {/* Padding bottom for better scrolling */}
-              <View className="h-4" />
-            </ScrollView>
-          </View>
-        </View>
-      </View>
-    </Modal>
-  );
-};
 
 const MemberMenu = ({ visible, onClose, isAdmin = false }: MemberMenuProps) => {
   if (!visible) return null;
@@ -383,42 +262,24 @@ export default function GroupInfo({ group }: GroupInfoProps) {
             </View>
           </View>
         ))}
-
-        {/* Add Member Button */}
-        <TouchableOpacity
-          className="flex-row items-center py-2 mt-2"
-          onPress={() => setShowAddMember(true)}
-        >
-          <View className="p-2 rounded-full bg-blue-50 items-center justify-center">
-            <Ionicons name="person-add-outline" size={16} color="#3B82F6" />
-          </View>
-          <Text className="ml-3 text-[15px] font-medium text-blue-500">
-            Thêm thành viên
-          </Text>
-        </TouchableOpacity>
       </View>
     );
   };
 
   return (
-    <View className="flex-1 px-4 pt-6 pb-4">
+    <View className="flex-1 px-4 pt-6 pb-3 border-b-4 border-gray-200">
       <View className="flex-row justify-between items-center mb-4">
         <Text className="text-base font-medium text-blue-950">
           Thành viên nhóm ({MOCK_MEMBERS.length})
         </Text>
         <TouchableOpacity
-          className="py-1 px-3 rounded-lg bg-blue-50 active:bg-blue-100"
+          className="py-1 px-3"
           onPress={() => setShowDetail(!showDetail)}
         >
-          <Text className="text-blue-500">Xem tất cả</Text>
+          <Ionicons name={showDetail ? "chevron-up" : "chevron-down"} size={18} color="#3B82F6" />
         </TouchableOpacity>
       </View>
       {showDetail && <ShowDetail />}
-
-      <AddMemberModal
-        visible={showAddMember}
-        onClose={() => setShowAddMember(false)}
-      />
     </View>
   );
 }
